+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 * warning, breaking changes were introduced:
     * the order of the parameters in CTMDataset was changed (now first is contextual embeddings)
     * CTM takes in input bow_size, contextual_size instead of input_size and bert_size
+    * changed the name of the parameters in the dataset
 * introduced early stopping
 
 1.8.2 (2021-02-08)